[fix] comment count icon 에서 숫자 제거
https://github.com/TeamCrazyPerformance/tcp-web/pull/10#discussion_r396358942
