--- src/components/atoms/CommentCountIcon/style.scss
@@ -1,12 +1,0 @@
-.comment_count_wrapper {
-    & {
-        line-height: 2.5rem;
-        text-align: center;
-        font-weight: lighter;
-        font-size: 0.8rem;
-    }
-
-    & > img {
-        vertical-align: middle;
-    }
-}
